Añadidos cambios a controlador y vista para la historia 3 poder añadir productos a mi carrito de compras.
