+++ CI0126-ExamenFinal-B70866/CI0126-ExamenFinal-B70866.csproj
@@ -481,7 +481,6 @@
     <Compile Include="App_Start\RouteConfig.cs" />
     <Compile Include="Controllers\HomeController.cs" />
     <Compile Include="Controllers\ProductController.cs" />
-    <Compile Include="Controllers\ShoppingCartController.cs" />
     <Compile Include="Global.asax.cs">
       <DependentUpon>Global.asax</DependentUpon>
     </Compile>
@@ -533,7 +532,7 @@
     <Content Include="Views\Home\Index.cshtml" />
     <Content Include="Views\Product\addProduct.cshtml" />
     <Content Include="Views\Product\productList.cshtml" />
-    <Content Include="Views\ShoppingCart\shoppingCart.cshtml" />
+    <Content Include="Views\Product\shoppingCart.cshtml" />
   </ItemGroup>
   <ItemGroup>
     <Folder Include="App_Data\" />
