+++ CI0126-ExamenFinal-B70866/SQL Scripts/ShoppingCart.sql
@@ -1,5 +1,5 @@
 ﻿CREATE TABLE ShoppingCart(
-	productID INTEGER PRIMARY KEY IDENTITY,
+	productID INTEGER,
 	productAmount INTEGER NOT NULL,
 	FOREIGN KEY(productID) REFERENCES Product(productID)
 )
