+++ CI0126-ExamenFinal-B70866/Views/Product/productList.cshtml
@@ -29,7 +29,7 @@
                 <img src=@Url.Action("fileAccess", "Product", new {productID = product.id}) style="width:128px;height:128px;" />
             </div>
             <div>
-                @Html.ActionLink("Añadir al carrito", "addProductToCart", "ShoppingCart", new { newProduct = product, amount = 1 }, new { @class = "btn btn-block light-blue lighten-4 black-text" })
+                @Html.ActionLink("Añadir al carrito", "addProductToCart", new { productID = product.id, amount = 1 })
             </div>
         }
     }
