+++ CI0126-ExamenFinal-B70866/Views/Product/shoppingCart.cshtml
@@ -16,7 +16,10 @@
             <div>
                 @product.amount
             </div>
+            <br />
+            <br />
         }
     }
 </body>
 </html>
+
