+++ CI0126-ExamenFinal-B70866/Views/Shared/_Layout.cshtml
@@ -25,7 +25,7 @@
                     <li>@Html.ActionLink("Contact", "Contact", "Home")</li>
                     <li>@Html.ActionLink("Añadir Producto", "addProduct", "Product")</li>
                     <li>@Html.ActionLink("Ver Productos", "productList", "Product")</li>
-                    <li>@Html.ActionLink("Carrito de Compras", "shoppingCart", "ShoppingCart")</li>
+                    <li>@Html.ActionLink("Carrito de Compras", "shoppingCart", "Product")</li>
                 </ul>
             </div>
         </div>
